+++ sample/test-fns/jest.config.cjs
@@ -1,4 +1,0 @@
-// sample/test-fns/jest.config.cjs
-
-module.exports = { ...require('../jest.config.default.cjs'),
-};