+++ sample/test-fns/jest.config.js
@@ -1,0 +1,4 @@
+// sample/test-fns/jest.config.js
+
+import opts from '../jest.config.default.js'
+export default opts