+++ sample/test-rules/jest.config.cjs
@@ -1,6 +1,0 @@
-// sample/test-rules/jest.config.cjs
-
-module.exports = { ...require('../jest.config.default.cjs'),
-  // Load docs, once at the beginning of the tests.
-  globalSetup: "./setup.jest.cjs"
-};